반응형 구현
