+++ components/BestListForm.module.css
@@ -1,11 +1,10 @@
 .BestList-form {
-  width: 384px;
+  width: 100%;
   height: 169px;
   border-radius: 8px;
   border: none;
   background-color: #f9fafb;
   margin-top: 24px;
-  padding: 0 24px;
   display: grid;
 }
 .BestList-badge {
@@ -28,7 +27,7 @@
   color: #ffffff;
 }
 .BestList-wrapper-h3 {
-  width: 336px;
+  width: 100%;
   height: 72px;
   display: flex;
   justify-content: space-between;
@@ -79,4 +78,34 @@
 }
 .date {
   color: #9ca3af;
+}
+@media screen and (min-width:1200px) {
+  .BestList-form{
+    padding: 0 24px;
+  }
+}
+@media screen and (min-width: 768px) and (max-width: 1199px) {
+  .BestList-badge{
+    margin-left:24px;
+  }
+  .BestList-wrapper-h3{
+    margin: 0 24px;
+  }
+  .BestList-wrapper{
+    margin: 0 24px;
+    margin-bottom: 16px;
+  }
+}
+@media screen and (min-width: 375px) and (max-width: 767px) {
+  .BestList-badge{
+    margin-left:24px;
+  }
+  .BestList-wrapper-h3{
+    margin: 0 24px;
+  }
+  .BestList-wrapper{
+    margin: 0 24px;
+    margin-bottom: 16px;
+  }
+
 }
