+++ components/Dropwdown.module.css
@@ -25,3 +25,14 @@
   line-height: 24px;
   color: #1f2937;
 }
+@media screen and (min-width: 375px) and (max-width: 767px) {
+  .Dropdown-select{
+    width:24px;
+    height:24px;
+    background-image: url('../public/sort.svg');
+    appearance: none;
+    background-repeat: no-repeat;
+    background-position: center;
+    text-indent:-999px;
+  }
+}
